parametrize script

+++ tools/dockerwithdata/createImage.sh
@@ -1,6 +1,11 @@
 #!/bin/bash
 set -euox pipefail
 
+if [ "$#" -ne 3 ]; then
+	echo "use $0 <gcloud-zoneid> <gcloud-vmname> <gcloud-project>"
+	exit -1
+fi
+
 docker stop tablelandbuildimage || true; docker rm tablelandbuildimage || true;
 DOCKER_TAG=$(date +%Y%m%d_%H%M%S)
 
@@ -8,16 +13,16 @@ docker build -f ../../local/postgres.Dockerfile -t tableland/postgres:$DOCKER_TA
 docker run -d --name tablelandbuildimage -e POSTGRES_USER=admin -e POSTGRES_PASSWORD=admin -e PGDATA=/data tableland/postgres:$DOCKER_TAG
 
 gcloud compute ssh \
---zone "us-west1-b" "tableland-postgres-production"  \
---project "textile-310716" \
+--zone $1 $2  \
+--project $3 \
 --command="sudo su postgres -c 'pg_dumpall --clean --if-exists'" > ../../dump.sql
 
 until docker exec tablelandbuildimage pg_isready -U admin; do sleep 3; done
 docker exec -i tablelandbuildimage psql -U admin < ../../dump.sql
 rm ../../dump.sql
 
-docker container commit tablelandbuildimage tableland/postgres:$DOCKER_TAG
+docker container commit tablelandbuildimage textile/tableland-postgres:$DOCKER_TAG
 docker stop tablelandbuildimage
 docker rm tablelandbuildimage
 
-echo "Successfully created Docker image tableland/postgres:$DOCKER_TAG"
+echo "Successfully created Docker image textile/tableland-postgres:$DOCKER_TAG"
